+++ src/demo/app/app.module.ts
@@ -1,11 +1,11 @@
 import { NgModule }      from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { AsyncLocalStorageModule } from 'angular-async-local-storage';
+import { LocalStorageModule } from '@ngx-pwa/local-storage';
 
 import { AppComponent }  from './app.component';
 
 @NgModule({
-  imports:      [ BrowserModule, AsyncLocalStorageModule ],
+  imports:      [ BrowserModule, LocalStorageModule ],
   declarations: [ AppComponent ],
   bootstrap:    [ AppComponent ]
 })