+++ src/lib/src/service/databases/localstorage-database.ts
@@ -4,10 +4,10 @@ import { map } from 'rxjs/operators';
 import { of as observableOf } from 'rxjs/observable/of';
 import { _throw as observableThrow } from 'rxjs/observable/throw';
 
-import { AsyncLocalDatabase } from './async-local-database';
+import { LocalDatabase } from './local-database';
 
 @Injectable()
-export class LocalStorageDatabase extends AsyncLocalDatabase {
+export class LocalStorageDatabase extends LocalDatabase {
 
   /* Initializing native localStorage right now to be able to check its support on class instanciation */
   protected localStorage = localStorage;